программа для 3д отрезка

+++ Task21/Program.cs
@@ -18,9 +18,9 @@ int Coordinate(string coorName, string pointName)
     return Convert.ToInt32(Console.ReadLine());
 }
 
-double Decision(double x1, double x2, 
-                double y1, double y2, 
-                double z1, double z2){
+double Decision(int x1, int x2, 
+                int y1, int y2, 
+                int z1, int z2){
   return Math.Sqrt(Math.Pow((x2-x1), 2) + 
                    Math.Pow((y2-y1), 2) + 
                    Math.Pow((z2-z1), 2));
